+++ Source Files/Backend/TripleDES.java
@@ -1,7 +1,7 @@
 package Backend;
 
 public class TripleDES extends Encryption{
-    public TripleDES(int keyIn, String textIn){
+    public TripleDES(float keyIn, String textIn){
         encryptMethod = "3DES";
         text = textIn;
         key = keyIn;
